refactor: Move alert dialog components to separate dir
