+++ src/components/Molucules/AlertDialog/AlerDialogPortal/AlerDialogPortal.component.tsx
@@ -1,5 +1,5 @@
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../../lib/utils';
 
 const AlertDialogPortal = ({
   className,
