+++ src/components/Molucules/AlertDialog/AlertDialogDescription/AlertDialogDescription.component.tsx
@@ -1,6 +1,6 @@
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
 import { forwardRef } from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../../lib/utils';
 
 const AlertDialogDescription = forwardRef<
   React.ElementRef<typeof AlertDialogPrimitive.Description>,
