+++ src/components/Molucules/AlertDialog/AlertDialogFooter/AlertDialogFooter.component.tsx
@@ -1,5 +1,5 @@
 import { HTMLAttributes } from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../../lib/utils';
 
 const AlertDialogFooter = ({
   className,
