+++ src/components/Molucules/AlertDialog/AlertDialogHeader/AlertDialogHeader.component.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../../lib/utils';
+import { cn } from "../../../../lib/utils";
 
 const AlertDialogHeader = ({
   className,
