+++ src/components/Molucules/AlertDialog/AlertDialogTitle/AlertDialogTitle.component.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from 'react';
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../../lib/utils';
 
 const AlertDialogTitle = forwardRef<
   React.ElementRef<typeof AlertDialogPrimitive.Title>,
